refactor(cli-core): move connection lookups into their resource modules
connectivity.ts was a shared module in name only: it held four unrelated,
use-case-specific lookups that each belong to an existing resource domain.
Moved them next to their siblings and deleted the module.

- getProjectSshConnection -> resources/project.ts
- getAppSshConnection     -> resources/app.ts
- getMysqlConnection, getPhpMyAdminUrl -> resources/database.ts
- getBackupDownloadUrl    -> resources/infrastructure.ts

Each now uses its module's existing error-handling convention, and the
duplicated main-user lookup shared by getMysqlConnection and
getPhpMyAdminUrl is a single helper in database.ts. No behaviour or public
function names change.
