+++ packages/mittwald-cli-core/src/resources/infrastructure.ts
@@ -245,6 +245,101 @@ export async function deleteBackup(options: DeleteBackupOptions): Promise<Librar
   return executeApiCall(options.apiToken, (client) => client.backup.deleteProjectBackup({ projectBackupId: options.backupId }), 204);
 }
 
+export interface GetBackupDownloadUrlOptions extends LibraryFunctionBase {
+  backupId: string;
+  /** Archive format; only applied when a new export is created (default: tar) */
+  format?: 'tar' | 'zip';
+  /** Password to protect the archive with; only applied when a new export is created */
+  password?: string;
+  /** Request a new export even if a usable one already exists */
+  recreate?: boolean;
+}
+
+export interface BackupDownloadUrl {
+  backupId: string;
+  /** Set once the export has finished; undefined while it is still being prepared */
+  downloadURL?: string;
+  format?: string;
+  phase?: string;
+  expiresAt?: string;
+  sha256Checksum?: string;
+  withPassword?: boolean;
+  /** True when this call triggered a new export */
+  exportRequested: boolean;
+}
+
+/**
+ * Returns the download URL of a backup's export, requesting a new export when none
+ * usable exists.
+ *
+ * Exports are prepared asynchronously, so a freshly requested one has no URL yet;
+ * callers poll by calling this again.
+ */
+export async function getBackupDownloadUrl(
+  options: GetBackupDownloadUrlOptions
+): Promise<LibraryResult<BackupDownloadUrl>> {
+  const startTime = performance.now();
+
+  try {
+    const client = MittwaldAPIV2Client.newWithToken(options.apiToken);
+
+    const backupResponse = await client.backup.getProjectBackup({ projectBackupId: options.backupId });
+    assertStatus(backupResponse, 200);
+
+    const format = options.format ?? 'tar';
+    const existing = backupResponse.data.export;
+    const usable =
+      existing &&
+      existing.format === format &&
+      existing.phase !== 'Failed' &&
+      existing.phase !== 'Expired';
+
+    if (usable && !options.recreate) {
+      return {
+        data: {
+          backupId: options.backupId,
+          downloadURL: existing.downloadURL,
+          format: existing.format,
+          phase: existing.phase,
+          expiresAt: existing.expiresAt,
+          sha256Checksum: existing.sha256Checksum,
+          withPassword: existing.withPassword,
+          exportRequested: false,
+        },
+        status: backupResponse.status,
+        durationMs: performance.now() - startTime,
+      };
+    }
+
+    const exportResponse = await client.backup.createProjectBackupExport({
+      projectBackupId: options.backupId,
+      data: {
+        format,
+        ...(options.password ? { password: options.password } : {}),
+      },
+    });
+    assertStatus(exportResponse, 204);
+
+    return {
+      data: {
+        backupId: options.backupId,
+        format,
+        phase: 'Pending',
+        withPassword: Boolean(options.password),
+        exportRequested: true,
+      },
+      status: exportResponse.status,
+      durationMs: performance.now() - startTime,
+    };
+  } catch (error) {
+    throw new LibraryError(
+      error instanceof Error ? error.message : 'Unknown error',
+      (error as any).status || 500,
+      { originalError: error, durationMs: performance.now() - startTime }
+    );
+  }
+}
+
 // Backup schedules
 export interface ListBackupSchedulesOptions extends LibraryFunctionBase {
   projectId: string;
